Move parse_texte tests to a dedicated place
